Moved the schemas to core and updated the imports
