+++ app/core/schemas/__init__.py
@@ -1,0 +1,19 @@
+from .datamodule import (
+    BaseDataModule,
+    DEFAULT_BATCH_SIZE,
+    DEFAULT_NUM_WORKERS,
+    DEFAULT_VAL_SPLIT,
+    DEFAULT_TEST_SPLIT,
+    DEFAULT_IMAGE_CLASSIFIER_HF_ID,
+    logger,
+)
+
+__all__ = [
+    "BaseDataModule",
+    "DEFAULT_BATCH_SIZE",
+    "DEFAULT_NUM_WORKERS",
+    "DEFAULT_VAL_SPLIT",
+    "DEFAULT_TEST_SPLIT",
+    "DEFAULT_IMAGE_CLASSIFIER_HF_ID",
+    "logger",
+]
