+++ app/vision_automl/router.py
@@ -12,7 +12,7 @@ from fastapi import APIRouter, Form, Request
 from fastapi.responses import JSONResponse
 
 from app.core.exceptions import AutoDWDownloadError, AutoMLValidationError
-from app.vision_automl.schemas.ml_tasks import SUPPORTED_VISION_TASK_TYPES
+from app.core.schemas.ml_tasks import SUPPORTED_VISION_TASK_TYPES
 from app.vision_automl.services import (
     build_upload_payload,
     convert_leaderboard_safely,
